requirements.txt
torch
torchvision
requests
gdown
matplotlib
opencv-python
Pillow<7
scikit-image
gradio
